[drm,lkpi] import headers from 4.11

+++ include/drm/intel-gtt.h
@@ -9,8 +9,10 @@ struct agp_bridge_data;
 struct intel_gtt;
 struct intel_gtt *intel_gtt_get(void);
 #else
-void intel_gtt_get(u64 *gtt_total, size_t *stolen_size,
-		   phys_addr_t *mappable_base, u64 *mappable_end);
+void intel_gtt_get(u64 *gtt_total,
+		   u32 *stolen_size,
+		   phys_addr_t *mappable_base,
+		   u64 *mappable_end);
 #endif
 
 int intel_gmch_probe(struct pci_dev *bridge_pdev, struct pci_dev *gpu_pdev,
